perfet the propetry between player and monster
*) make some balance for propetry
*) add critical logic
*) add feathure of (indicate monster rank from xml)
*) the monster's level can be modify 10% of src propetry

+++ app/src/main/java/com/cb/adventures/data/MonsterPropetry.java
@@ -68,7 +68,7 @@ public class MonsterPropetry extends Propetry implements Cloneable{
         this.rightFrames = rightFrames;
     }
 
-    public int getSpeed() {
+    public float getSpeed() {
         return speed;
     }
